task1/CTPN Method/requirements.txt
torch>=2.5,<2.7
torchvision>=0.20,<0.22
pillow>=9,<12
numpy>=1.24,<3
opencv-python>=4.8,<5
